feat #39 : 사용자 정보 조회 및 닉네임 변경 api 연동
