+++ src/infrastructure/api/user/index.ts
@@ -1,5 +1,17 @@
 import client from "@api";
 
+export const getMyInfo = async () =>
+  await client
+    .get("/members")
+    .then(({ data }) => data)
+    .catch((err) => err.response);
+
+export const editUserNickName = async (newNickname: string) =>
+  await client
+    .put("/members/nicknames", { newNickname })
+    .then(({ data }) => data)
+    .catch((err) => err.response);
+
 export const logout = async () =>
   await client
     .delete("/auth/logout")
